refactor adpost screen
