+++ lib/features/post_ad/presentation/widgets/post_ad_app_bar.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:skillbridge/core/routing/app_navigator.dart';
+import 'package:skillbridge/core/theme/app_colors.dart';
+import 'package:skillbridge/core/theme/app_styles.dart';
+
+class PostAdAppBar extends StatelessWidget implements PreferredSizeWidget {
+  @override
+  final Size preferredSize = const Size.fromHeight(kToolbarHeight + 1);
+
+  const PostAdAppBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      backgroundColor: AppColors.white,
+      elevation: 0,
+      scrolledUnderElevation: 0,
+      centerTitle: true,
+      leading: GestureDetector(
+        onTap: () => context.pop(),
+        child: Container(
+          margin: EdgeInsets.all(10.w),
+          decoration: BoxDecoration(
+            border: Border.all(color: AppColors.border),
+            borderRadius: BorderRadius.circular(10.r),
+          ),
+          child: Icon(Icons.arrow_back, size: 18.sp, color: AppColors.textDark),
+        ),
+      ),
+      title: const Text('Post an Ad', style: AppStyles.font17Bold),
+      bottom: PreferredSize(
+        preferredSize: const Size.fromHeight(1),
+        child: Container(height: 1, color: AppColors.border),
+      ),
+    );
+  }
+}
